nxp-s32g: README: update to compatible with pkcs11-hse recipe
Issue: LINUXEXEC-26603

Because the settings to compile pkcs11-hse have been changed, so update
the README to compatible with it.

In simply, we introduced new variables
    HSE_FW_VERSION_S32G2, HSE_FW_VERSION_S32G3
to indicate the HSE firmware version to use, to make it clear and covenient
for users to use HSE feature.

By adding the new variables, users can specify the HSE firmware version to use.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>