nxp-s32g: hse: fix the default HSE firmware version to v0.1.0.5
Issue: LINUXEXEC-26603

The v0.1.0.9 HSE firmware will affect S32G secure boot function.
And the default version of HSE firmware is v0.1.0.5 too, in the SDK BSP36
project settings, so update README and conf/machine/include/hse.inc to
use v0.1.0.5, althouth the latest v0.1.0.9 support description is included
in BSP36 user manual.

And there is a new created ticket to NXP to track the status of this issue:
  https://community.nxp.com/t5/S32G/RDB2-boot-failed-with-HSE-FW-v0-1-0-9-based-on-BSP36/m-p/1654097#M3282

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>